Add Allauth and ContactInfo for CustomUser
Note that this branch is not up to date because:
The Room -> Unit change in the latest branch is not fully working.
Therefore, I used the version of 5 days ago (similar to the current master branch).

Please run `pip install -r requirements.txt` in your environment.
If you cannot make migrations, please delete estate/migrations/0xxx.py and estate/migrations/__pycache__.
Please follow this link https://learndjango.com/tutorials/django-allauth-tutorial and skip to the Github OAuth part to configure the application at the admin page for the GitHub authentication.